Extracting the changelog string rendering into separate classes.
