+++ src/Generator/Changelog/Changesets/RepresentationData.php
@@ -1,0 +1,52 @@
+<?php
+namespace Mill\Generator\Changelog\Changesets;
+
+use Mill\Generator\Changelog\Changeset;
+
+class RepresentationData extends Changeset
+{
+    /**
+     * @var array
+     */
+    protected $templates = [
+        'plural' => [
+            'added' => 'The {representation} has added the following fields:',
+            'removed' => 'The {representation} has removed the following fields:'
+        ],
+        'singular' => [
+            'added' => '{field} has been added to the {representation} representation.',
+            'removed' => '{field} has been removed from the {representation} representation.'
+        ]
+    ];
+
+    /**
+     * @inheritdoc
+     */
+    public function compileAddedOrRemovedChangeset($definition, array $changes = [])
+    {
+        if (count($changes) === 1) {
+            $change = array_shift($changes);
+            $template = $this->templates['singular'][$definition];
+            return $this->renderText($template, $change);
+        }
+
+        $fields = [];
+        foreach ($changes as $change) {
+            $fields[] = $this->renderText('{field}', $change);
+        }
+
+        $template = $this->templates['plural'][$definition];
+        return [
+            $this->renderText($template, array_shift($changes)),
+            $fields
+        ];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function compileChangedChangeset($definition, array $changes = [])
+    {
+        throw new \Exception($definition . ' representation data changes are not yet supported.');
+    }
+}
